HBASE-24332: TestJMXListener.setupBeforeClass can fail due to not getting a random port.
