+++ hbase-http/src/test/java/org/apache/hadoop/hbase/http/TestProxyUserSpnegoHttpServer.java
@@ -170,7 +170,7 @@ public class TestProxyUserSpnegoHttpServer extends HttpServerFunctionalTest {
     kdc.setWorkDir(kdcDir);
 
     kdc.setKdcHost(KDC_SERVER_HOST);
-    int kdcPort = getFreePort();
+    int kdcPort = HBaseCommonTestingUtility.randomFreePort();
     kdc.setAllowTcp(true);
     kdc.setAllowUdp(false);
     kdc.setKdcTcpPort(kdcPort);
